feat(ObjTriplet): make mutable
i made the library to have almost zero encapsulation, because the
classes should act as data transfer objects. but not immutable DTO's,
but as a "database" for the user. library parses the file to give to the
user the output and forget about it. encapsulation between all the
different entities will be too difficult and too restrictive, so only
nulls and obvious checks are evaluated.